requirements.txt
# packages in environment at C:\Users\Griffin\anaconda3\envs\pyOrbital:
#
# Name                    Version                   Build  Channel
ca-certificates           2021.5.30            h5b45459_0    conda-forge
certifi                   2021.5.30        py39hcbf5309_0    conda-forge
cycler                    0.10.0           py39haa95532_0  
freetype                  2.10.4               hd328e21_0  
geos                      3.9.1                h39d44d4_2    conda-forge
intel-openmp              2021.3.0          haa95532_3372  
jpeg                      9b                   hb83a4c4_2  
kiwisolver                1.3.1            py39hd77b12b_0  
matplotlib-base           3.4.2            py39h49ac443_0  
mkl-service               2.4.0            py39h2bbff1b_0  
munkres                   1.1.4                      py_0  
numpy-base                1.20.3           py39hc2deb75_0  
olefile                   0.46                       py_0  
openssl                   1.1.1k               h8ffe710_1    conda-forge
python-dateutil           2.8.2              pyhd3eb1b0_0  
setuptools                52.0.0           py39haa95532_0  
shapely                   1.7.1            py39haadaec5_5    conda-forge
sqlite                    3.36.0               h2bbff1b_0  
vs2015_runtime            14.27.29016          h5e58377_2  
wheel                     0.37.0             pyhd3eb1b0_0  
wincertstore              0.2              py39h2bbff1b_0  
